rename package naming.net --> naming.remote
